requirements.txt
google-generativeai==0.3.0
python-dotenv==1.0.0
reportlab==4.0.4
python-docx==0.8.11
